WinRT: added project files for Windows 8.0/RT use
The copy of Freetype from "externals/" was used, however no modifications were
made to it.  If this copy of Freetype is updated, the paths inside the
project files (in VisualC-WinRT) will need to be updated.

Freetype was linked to SDL_ttf via static linking, as it's required for use
by SDL_ttf, and helped prevent Freetype from being modified.

TODO:
- write a script to update project file contents if and when an external library
  is updated.
- add project files for Windows Phone 8.0
- add project files for WinRT Universal (Windows 8.1, Windows Phone 8.1, etc.)
- add a README.txt file to VisualC-WinRT/, outlining which directories are
  relevant to which project types (Windows 8.0/RT has one project type,
  Windows Phone 8.0 has another, WinRT universal has another)