Fixed bug 4343 - style Outline overlap
Sylvain

When using Outline style with a big value (~50). Text starts to overlap (the end being written at the beginning.

because the following check is un-activated with an outline not 0.

        /* Ensure the width of the pixmap is correct. On some cases,
         * freetype may report a larger pixmap than possible.*/
        width = current->width;
        if (font->outline <= 0 && width > glyph->maxx - glyph->minx) {
            width = glyph->maxx - glyph->minx;
        }

That was done with initial implementation because the maxx/minx were not updated with outline. Since https://hg.libsdl.org/SDL_ttf/rev/ff81ac6d76c4 it is. ( cached->maxx += 2 * fo; )
Now that we can reactivated this sanity check.

Moreover, the fact we hit the overlap means that the pixmap/bitmap is cropped a little bit. We can increase a little bit the bounding box..